+++ packages/waggle/src/registration/registrationService.test.ts
@@ -44,7 +44,7 @@ afterEach(async () => {
 })
 
 async function registerUserTest(csr: string): Promise<Response> {
-  const socksProxyAgent = new SocksProxyAgent({ port: ports.socksPort, host: 'localhost', timeout: 60000 })
+  const socksProxyAgent = new SocksProxyAgent({ port: ports.socksPort, host: 'localhost', timeout: 100000 })
   const options = {
     method: 'POST',
     body: JSON.stringify({ data: csr }),