+++ packages/waggle/src/socket/constantsReponse.ts
@@ -9,5 +9,7 @@ export enum EventTypesResponse {
   RESPONSE_FETCH_ALL_DIRECT_MESSAGES = 'responseFetchAllDirectMessages',
   SEND_PEER_ID = 'sendPeerId',
   SEND_IDS = 'sendIds',
+  SEND_USER_CERTIFICATE = 'sendUserCertificate',
+  CERTIFICATE_REGISTRATION_ERROR = 'certificateRegistrationError',
   RESPONSE_GET_CERTIFICATES = 'responseGetCertificates'
 }