+++ packages/waggle/src/socket/listeners/connection.ts
@@ -62,6 +62,9 @@ export const connections = (io, connectionsManager: ConnectionsManager) => {
         await connectionsManager.askForMessages(channelAddress, ids)
       }
     )
+    socket.on(EventTypesServer.REGISTER_USER_CERTIFICATE, async (serviceAddress: string, userCsr: string) => {
+      await connectionsManager.registerUserCertificate(serviceAddress, userCsr)
+    })
     socket.on(EventTypesServer.SAVE_CERTIFICATE, async (certificate: string) => {
       console.log('Received saveCertificate websocket event, processing.')
       await connectionsManager.saveCertificate(certificate)